riscv/cfi: Enable CFI on static binaries
For static binaries, CFI are enabled inside ARCH_SETUP_TLS, with a macro
to enable the shadow stack to prevent underflowing the shadow stack on
return, and with a function _dl_cfi_setup_features to enable landing
pad.

It scans backward of the program header to find the PT_GNU_PROPERTY note
first, then enable CFI features corresponding to the feature bits.

Co-authored-by: Deepak Gupta <debug@rivosinc.com>