riscv/cfi: Store shadow stack information in TLS
Switching between different ucontexts involves two cases.  One case is
when both contexts share the same shadow stack, which requires unwinding.
The other is when each context uses a different shadow stack, which
requires a stack switch using a shadow stack restore token.

By storing the shadow stack base in TLS and comparing it to the value
stored in the context, we can tell the difference and perform the right
action.